App/requirements.txt
boardgamegeek2==1.0.1
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-crispy-forms==1.7.2
django-registration-redux==2.6
django-widget-tweaks==1.4.3
djangorestframework==3.9.2
idna==2.8
psycopg2==2.7.7
pytz==2018.9
requests==2.21.0
requests-cache==0.4.13
urllib3==1.24.1
wincertstore==0.2
